Move functions from __init__ to backend in preparation of split
