+++ tests/test_init.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import pytest
 
 import mkosi
+import mkosi.backend
 
 
 def test_fedora_release_cmp() -> None:
@@ -30,6 +31,7 @@ def test_strip_suffixes() -> None:
     assert mkosi.strip_suffixes(Path("home.xz/test")) == Path("home.xz/test")
     assert mkosi.strip_suffixes(Path("home.xz/test.txt")) == Path("home.xz/test.txt")
 
+
 def test_copy_file(tmpdir: Path) -> None:
         dir_path = Path(tmpdir)
         file_1 = Path(dir_path) / "file_1.txt"
@@ -38,19 +40,19 @@ def test_copy_file(tmpdir: Path) -> None:
         file_2.touch()
 
         # Copying two empty files.
-        mkosi.copy_file(file_1, file_2)
+        mkosi.backend.copy_file(file_1, file_2)
         assert filecmp.cmp(file_1, file_2)
 
         # Copying content from one file.
         file_1.write_text("Testing copying content from this file to file_2.")
-        mkosi.copy_file(file_1, file_2)
+        mkosi.backend.copy_file(file_1, file_2)
         assert filecmp.cmp(file_1, file_2)
 
         # Giving a non existing path/file.
         with pytest.raises(OSError):
-            mkosi.copy_file("nullFilePath", file_1)
+            mkosi.backend.copy_file("nullFilePath", file_1)
 
         # Copying when there's already content in both files.
         file_2.write_text("Testing copying content from file_1 to file_2, with previous data.")
-        mkosi.copy_file(file_1, file_2)
+        mkosi.backend.copy_file(file_1, file_2)
         assert filecmp.cmp(file_1, file_2)
